+++ Panther/include/AST/ASTBuffer.h
@@ -344,6 +344,18 @@ namespace pcit::panther{
 				return this->news[node._value.node_index];
 			}
 
+			EVO_NODISCARD auto createArrayInitNew(auto&&... args) -> AST::Node {
+				evo::debugAssert(this->is_locked == false, "Cannot create as buffer is locked");
+				const uint32_t node_index = this->array_init_news.emplace_back(std::forward<decltype(args)>(args)...);
+				return AST::Node(AST::Kind::ARRAY_INIT_NEW, node_index);
+			}
+			EVO_NODISCARD auto getArrayInitNew(const AST::Node& node) const -> const AST::ArrayInitNew& {
+				evo::debugAssert(
+					node.kind() == AST::Kind::ARRAY_INIT_NEW, "Node is not a ArrayInitNew"
+				);
+				return this->array_init_news[node._value.node_index];
+			}
+
 			EVO_NODISCARD auto createStructInitNew(auto&&... args) -> AST::Node {
 				evo::debugAssert(this->is_locked == false, "Cannot create as buffer is locked");
 				const uint32_t node_index = this->struct_init_news.emplace_back(std::forward<decltype(args)>(args)...);
@@ -452,6 +464,7 @@ namespace pcit::panther{
 			core::LinearStepAlloc<AST::MultiAssign, uint32_t> multi_assigns{};
 
 			core::LinearStepAlloc<AST::New, uint32_t> news{};
+			core::LinearStepAlloc<AST::ArrayInitNew, uint32_t> array_init_news{};
 			core::LinearStepAlloc<AST::StructInitNew, uint32_t> struct_init_news{};
 			core::LinearStepAlloc<AST::TryElse, uint32_t> try_elses{};