+++ Panther/src/Diagnostic.cpp
@@ -62,6 +62,7 @@ namespace pcit::panther{
 			case AST::Kind::POSTFIX:          return Location::get(ast_buffer.getPostfix(node), src);
 			case AST::Kind::MULTI_ASSIGN:     return Location::get(ast_buffer.getMultiAssign(node), src);
 			case AST::Kind::NEW:              return Location::get(ast_buffer.getNew(node), src);
+			case AST::Kind::ARRAY_INIT_NEW:   return Location::get(ast_buffer.getArrayInitNew(node), src);
 			case AST::Kind::STRUCT_INIT_NEW:  return Location::get(ast_buffer.getStructInitNew(node), src);
 			case AST::Kind::TRY_ELSE:         return Location::get(ast_buffer.getTryElse(node), src);
 			case AST::Kind::TYPE_DEDUCER:     return Location::get(ast_buffer.getTypeDeducer(node), src);
@@ -194,6 +195,10 @@ namespace pcit::panther{
 	}
 
 	auto Diagnostic::Location::get(const AST::New& new_expr, const Source& src) -> Location {
+		return Location::get(new_expr.keyword, src);
+	}
+
+	auto Diagnostic::Location::get(const AST::ArrayInitNew& new_expr, const Source& src) -> Location {
 		return Location::get(new_expr.keyword, src);
 	}